[dev] added signup/login and also updated partials and main handlebar
