+++ package.json
@@ -26,6 +26,7 @@
     "express": "^4.18.2",
     "express-handlebars": "^7.0.4",
     "express-session": "^1.17.3",
+    "handlebars": "^4.7.7",
     "multer": "^1.4.5-lts.1",
     "mysql2": "^3.2.0",
     "sequelize": "^6.30.0"
