+++ public/js/logout.js
@@ -7,7 +7,7 @@ const logout = async () => {
   if (response.ok) {
     document.location.replace('/');
   } else {
-    alert('Failed to log out.');
+    alert('Failed to sign up.');
   }
 };
