+++ public/js/signup.js
@@ -1,0 +1,27 @@
+const signupFormHandler = async(event) => {
+    event.preventDefault();
+
+    const username = document.querySelector('#username-signup').value.trim();
+    const email = document.querySelector('#email-signup').value.trim();
+    const password = document.querySelector('#password-signup').value.trim();
+
+  
+    if (username && email && password) {
+        const response = await fetch('/api/user/signup', {
+            method: 'POST',
+            body: JSON.stringify({ username, email, password }),
+            headers: { 'Content-Type': 'application/json' },
+        })
+
+        if (response.ok) {
+            alert('You have signed up!')
+            document.location.replace('/login');
+        } else {
+            alert('Failed to sign up.');
+        }
+    }
+};
+
+document
+  .querySelector(".signup-form")
+  .addEventListener("submit", signupFormHandler);
