+++ views/layouts/main.handlebars
@@ -4,25 +4,33 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title></title>
-  <link rel="stylesheet" type="text/css" href="/css/style.css">
+
+  {{!-- bootstrap --}}
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet"
+    integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
+
+  {{!-- CSS --}}
+  {{!--
+  <link rel="stylesheet" type="text/css" href="/css/style.css"> --}}
+
+  <title>Share camGear</title>
 </head>
 
 <body>
-  <header>
-    <h1>Welcome to Rent My Stuff</h1>
-  </header>
-  <nav>
-    <a href="/">Home</a>
-  </nav>
-  <main>
-    <aside>
-      <h2>Rent your stuff here!</h2>
-    </aside>
-    <div class="gallery">
-      {{{body}}}
-    </div>
-  </main>
+  <div>
+    <header>
+      {{> navBar}}
+    </header>
+    <main>
+      {{{ body }}}
+    </main>
+    <footer>
+      {{> footer }}
+    </footer>
+  </div>
+  {{#if loggedIn}}
+  <script src="/js/logout.js"></script>
+  {{/if}}
 </body>
 
 </html>
