+++ views/login.handlebars
@@ -1,0 +1,31 @@
+<!--Login-->
+<div class="row">
+    <div class="w-50 col-md-8 mx-auto">
+        <h2 id="page-title">Login Page</h2>
+
+        <form class="form login-form">
+
+            <div class="form-group mt-3">
+                <label for="username-login">Username:</label>
+                <input class="form-input" type="text" id="username-login" />
+            </div>
+
+            <div class="form-group mt-3">
+                <label for="password-login">Password:</label>
+                <input class="form-input" type="password" id="password-login" />
+            </div>
+
+            <div class="form-group mt-3">
+                <button id="login" class="btn btn-primary" type="submit">Login</button>
+            </div>
+
+        </form>
+
+        <div>
+            <br>
+            <a href="/signup">Register here!</a>
+        </div>
+
+    </div>
+
+    <script src="./js/login.js"></script>
