+++ views/partials/footer.handlebars
@@ -1,3 +1,10 @@
-<footer>
-     <p>&copy; Copyright Group 5 </p>
-</footer>
+{{!-- This is the Footer partial. --}}
+<div class="text-center">
+     <p>
+          <strong>Assembled by: The Share camGear Team. </strong> The source code is
+          licensed
+          <a href="http://opensource.org/licenses/mit-license.php">MIT</a>.
+          The website content is powered by <a href="https://www.npmjs.com/package/multer">Multer</a>
+          & <a href="https://handlebarsjs.com/">Handlebars</a>.
+     </p>
+</div>
