+++ views/partials/navBar.handlebars
@@ -1,11 +1,35 @@
- <nav>
-      <div id="companyDetails">
-      <img src="" alt="logo">
-      <p>Company Name</p>
-      </div>
-      <ul>
-          <li><a href="">Home</a></li>
-          <li><a href="">My Profile</a></li>
-          <li><a href="">Login</a></li>
-      </ul>
-  </nav>
+{{!-- This is the navigation bar partial. --}}
+<nav class="navbar navbar-expand-md navbar-light bg-light">
+    <div class="container-fluid">
+        <div class="d-flex">
+            <div class="p-2">
+                <a class="navbar-brand" href="/">
+                    <img class="logo img-fluid" src="/images/cameraShare_logo.png" alt="logo"
+                        style="width: 200px; height: 200px;">
+                </a>
+            </div>
+            <div class="navbar-brand d-flex justify-content-center">
+                <h1>Share camGear</h1>
+            </div>
+        </div>
+        <div class="d-flex">
+            <ul class="navbar-nav ms-auto">
+                <li class="nav-item">
+                    <a class="nav-link" href="/">Home</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/myprofile">My Profile</a>
+                </li>
+                {{#if loggedIn}}
+                <li class="nav-item">
+                    <button id="logout" class="btn btn-primary" id="logout">Logout</button>
+                </li>
+                {{else}}
+                <li class="nav-item">
+                    <a id="login" class="btn btn-primary" href="/login">Login</a>
+                </li>
+                {{/if}}
+            </ul>
+        </div>
+    </div>
+</nav>
