+++ views/signUp.handlebars
@@ -1,0 +1,34 @@
+<!--Sign Up-->
+<div class="w-50 col-md-8 mx-auto">
+
+    <h2 id="page-title">Sign up</h2>
+
+    <form class="form signup-form">
+
+        <div class="form-group mt-3">
+            <label for="username-signup">Username:</label>
+            <input class="form-input" type="text" id="username-signup" />
+        </div>
+
+        <div class="form-group mt-3">
+            <label for="username-signup">Email:</label>
+            <input class="form-input" type="text" id="username-signup" />
+        </div>
+
+        <div class="form-group mt-3">
+            <label for="password-login">Password:</label>
+            <input class="form-input" type="password" id="password-signup" />
+            <small class="text-muted">Password must be at least 7 characters long.</small>
+        </div>
+
+        <div class="form-group mt-3">
+            <button id="login" class="btn btn-primary" type="submit">Register</button>
+        </div>
+
+    </form>
+
+</div>
+
+</div>
+
+<script src="./js/signup.js"></script>
